requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
mysqlclient==2.2.4
python-decouple==3.8
Pillow==10.2.0
django-filter==24.1
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
requests==2.31.0
dj-database-url==2.3.0
packaging==24.2
psycopg2-binary==2.9.10
typing_extensions==4.12.2
tzdata==2025.1
